refactor(core): promote MapManager to engine-level with collision support
- Moved MapManager, Tile, TileType from demo into core
- Introduced CollisionProvider interface for pluggable collision logic
- Updated PhysicsSystem to use CollisionProvider
- Cleaned up fallback behavior in DemoScene for map loading
- Prepared MapManager for tile loading and 2D rendering integration

This modularizes collision, makes 2D map logic engine-ready, and sets foundation for upcoming render backend swap (OpenGL or Java2D).

+++ src/main/java/com/arcadia/demo/EntityFactory.java
@@ -1,19 +1,19 @@
 package com.arcadia.demo;
 
 import com.arcadia.core.components.InputIntentComponent;
+import com.arcadia.core.components.PlayerControlledComponent;
 import com.arcadia.core.components.PositionComponent;
 import com.arcadia.core.components.RenderLayerComponent;
 import com.arcadia.core.components.RenderableComponent;
 import com.arcadia.core.components.VelocityComponent;
 import com.arcadia.core.entity.Entity;
 import com.arcadia.core.util.EngineLogger;
-import com.arcadia.demo.components.PlayerControlledComponent;
 
 public class EntityFactory {
 
     public static Entity createPlayer(double x, double y) {
     Entity player = new Entity();
-        player.addComponent(new PositionComponent(x, y));
+        player.addComponent(new PositionComponent((int)x, (int)y));
         player.addComponent(new VelocityComponent(0,0));
         player.addComponent(new InputIntentComponent());
         player.addComponent(new PlayerControlledComponent());
@@ -25,7 +25,7 @@ public class EntityFactory {
 
     public static Entity createWanderer(double x, double y) {
         Entity npc = new Entity();
-        npc.addComponent(new PositionComponent(x, y));
+        npc.addComponent(new PositionComponent((int)x, (int)y));
         npc.addComponent(new VelocityComponent(Math.random() - 0.5, Math.random() - 0.5));
         npc.addComponent(new RenderableComponent('E'));
         npc.addComponent(new RenderLayerComponent(1));
@@ -35,7 +35,7 @@ public class EntityFactory {
 
     public static Entity createStaticObject(double x, double y) {
         Entity obj = new Entity();
-        obj.addComponent(new PositionComponent(x, y));
+        obj.addComponent(new PositionComponent((int)x, (int)y));
         obj.addComponent(new RenderableComponent('O'));
         obj.addComponent(new RenderLayerComponent(0));    
         EngineLogger.spawn("Created Static Object: " + obj);
